Remove expired clients from the dashboard
Problem:
Since #1274, idle clients' metrics are reaped, but this expiry is
not reflected in the admin dashboard.

Solution:
* Remove expired clients' sections from the dashboard, and deregister
them from the MetricsCollector.
* In anticipation of having to do lots of lookups of handlers in the
MetricsCollector, I've changed the implementation to be map based
instead of searching through arrays to find handlers. (So more like
AddEventListener)
* Also add the concept of expired clients to the combined client graph
so that expired clients can be removed